Fixed issue with static directories, an extra / was appended to dirs that were aliased to the root directory. Example: nexus.get.route("./folder/", "/");
In this example, all items in ./folder/ became:
//item1
//item2
//item3

The solution is to allow staticDir() with only one argument, the folder.
This routes all items in the folder to the root directory.
If you do submit ("./folder/", "/") as arguments, it will use //
If this becomes an issue, staticDir can detect if alias is "/" and
ignore it.